backend/requirements.txt
celery>=5.2.1
flask>=2.0.2
flask-cors>=3.0.10
python-whois>=0.7.3
dnspython>=2.1.0